+++ integration/backward_compatibility.go
@@ -8,6 +8,7 @@ var DefaultPreviousVersionImages = map[string]func(map[string]string) map[string
 	"quay.io/cortexproject/cortex:v1.11.0": func(flags map[string]string) map[string]string {
 		flags["-store.engine"] = "blocks"
 		flags["-server.http-listen-port"] = "8080"
+		flags["-store-gateway.sharding-enabled"] = "true"
 		return flags
 	},
 }