+++ integration/integration_memberlist_single_binary_test.go
@@ -92,8 +92,8 @@ func testSingleBinaryEnv(t *testing.T, tlsEnabled bool, flags map[string]string)
 	require.NoError(t, mimir2.WaitSumMetrics(e2e.Equals(3), "memberlist_client_cluster_members_count"))
 	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(3), "memberlist_client_cluster_members_count"))
 
-	// Each Mimir instance will use (512 ingester tokens + 1 distributor token + 512 compactor tokens).
-	tokensPerInstance := float64(512 + 1 + 512)
+	// Each Mimir instance will use (512 ingester tokens + 1 distributor token + 512 compactor tokens + 512 store gateway tokens + 512 tokens seen by store-gateway client).
+	tokensPerInstance := float64(512 + 1 + 512 + 512 + 512)
 	// 3 = number of instances.
 	require.NoError(t, mimir1.WaitSumMetrics(e2e.Equals(3*tokensPerInstance), "cortex_ring_tokens_total"))
 	require.NoError(t, mimir2.WaitSumMetrics(e2e.Equals(3*tokensPerInstance), "cortex_ring_tokens_total"))
@@ -104,19 +104,20 @@ func testSingleBinaryEnv(t *testing.T, tlsEnabled bool, flags map[string]string)
 	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(0), "memberlist_client_kv_store_value_tombstones"))
 	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(0), "memberlist_client_kv_store_value_tombstones"))
 
-	// Tombstones should increase by three for each server that leaves (once for distributor, ingester and compactor ring)
+	// Tombstones should increase by four for each server that leaves (once for distributor, ingester, compactor and store-gateway ring)
+	tombstonesPerRemovedInstance := 4.0
 	require.NoError(t, s.Stop(mimir1))
 	require.NoError(t, mimir2.WaitSumMetrics(e2e.Equals(2*tokensPerInstance), "cortex_ring_tokens_total"))
 	require.NoError(t, mimir2.WaitSumMetrics(e2e.Equals(2), "memberlist_client_cluster_members_count"))
-	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(3), "memberlist_client_kv_store_value_tombstones"))
+	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(tombstonesPerRemovedInstance), "memberlist_client_kv_store_value_tombstones"))
 	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(2*tokensPerInstance), "cortex_ring_tokens_total"))
 	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(2), "memberlist_client_cluster_members_count"))
-	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(3), "memberlist_client_kv_store_value_tombstones"))
+	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(tombstonesPerRemovedInstance), "memberlist_client_kv_store_value_tombstones"))
 
 	require.NoError(t, s.Stop(mimir2))
 	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(1*tokensPerInstance), "cortex_ring_tokens_total"))
 	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(1), "memberlist_client_cluster_members_count"))
-	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(6), "memberlist_client_kv_store_value_tombstones"))
+	require.NoError(t, mimir3.WaitSumMetrics(e2e.Equals(2*tombstonesPerRemovedInstance), "memberlist_client_kv_store_value_tombstones"))
 
 	require.NoError(t, s.Stop(mimir3))
 }
@@ -195,8 +196,8 @@ func TestSingleBinaryWithMemberlistScaling(t *testing.T) {
 	// Sanity check the ring membership and give each instance time to see every other instance.
 
 	for _, c := range instances {
-		// we expect 3*maxMimir to account for both the ingester, distributor and compactor rings
-		require.NoError(t, c.WaitSumMetrics(e2e.Equals(float64(maxMimir*3)), "cortex_ring_members"))
+		// we expect 5*maxMimir to account for ingester, distributor, compactor, store-gateway and store-gateway-client rings
+		require.NoError(t, c.WaitSumMetrics(e2e.Equals(float64(maxMimir*5)), "cortex_ring_members"))
 		require.NoError(t, c.WaitSumMetrics(e2e.Equals(0), "memberlist_client_kv_store_value_tombstones"))
 	}
 
@@ -223,9 +224,9 @@ func TestSingleBinaryWithMemberlistScaling(t *testing.T) {
 	// We don't use WaitSumMetrics [over all instances] here so we can log the per-instance metrics.
 
 	// These values are multiplied by three to account for the fact that ingester
-	// distributor and compactor components use a ring.
-	expectedRingMembers := float64(minMimir) * 3
-	expectedTombstones := float64(maxMimir-minMimir) * 3
+	// distributor, compactor, store-gateway components use a ring.
+	expectedRingMembers := float64(minMimir) * 5 // One extra, because store-gateway-client uses ring too. But it's the same ring as store-gateway.
+	expectedTombstones := float64(maxMimir-minMimir) * 4
 
 	require.Eventually(t, func() bool {
 		ok := true